+++ README.md
@@ -5,7 +5,7 @@ colorFrom: green
 colorTo: yellow
 sdk: docker
 sdk_version: "4.22.0"
-app_file: yolo/app.py
+app_file: app.py
 pinned: false
 ---
