+++ yolo temp/app.py
@@ -1,0 +1,2 @@
+#redirector
+from yolo.app import app
